Make simple_message targets depend on generated headers in industrial_msgs.
Targets in simple_message cannot be build without the generated headers from
industrial_msgs. According to [1], the accepted solution is to declare an
explicit dependency (for those targets) on the EXPORTED_TARGETS [2] of the
package that contains the messages.

In this case, all targets in simple_message depend on those in industrial_msgs.

[1] https://groups.google.com/d/msg/ros-sig-buildsystem/dvVO5QCHBLM/8mDS2lQelyEJ
[2] http://ros.org/rosdoclite/groovy/api/catkin/html/dev_guide/generated_cmake_api.html#catkin-package