robot_client: workaround for #46. Fix #67.
This is an updated version of the workaround committed in 9df46977. Instead
of requiring dependent packages to invoke the function defined in the
CFG_EXTRAS cmake snippet, the snippet now sets up the linker path directly.

Dependent packages now only need to remember to explicitly list their
dependency on `industrial_robot_client` and `simple_message` in their
`add_library(..)` statements.