robot_simulator: add GetRobotInfo svc server.
Users can override the information returned by the get_robot_info service
server. This can be useful if multiple instances of the simulator are
used in a ROS application, and each needs to be uniquely identified.

Additionally, this makes it possible to return information taken from real
controllers and manipulators, in order to provide a more 'real' simulation.

Users can enable the override by setting the private parameter 'robot_info' in
a launch file like so:

  <node .. type="industrial_robot_simulator" name=".." ..>
    <rosparam param="robot_info">
      controller:
        model: 'My Robot Controller'
        serial_number: 'ABC-123'
        sw_version: 'V1.23/ABC'
      robots:
        - model: 'My Robot Model'
          serial_number: '123456'
          sw_version: 'V4.56/DEF'
    </rosparam>
  </node>

As the Industrial Robot Simulator currently simulates only a single robot /
motion group, specifying multiple robot entries will result in only the first
of the list to be used.

Fields absent from the rosparam parameter will result in empty strings being
assigned to the corresponding fields in the 'controller' and 'robot' parts of
the service response.