traj_action: abort/reject based on motion server state
The action server will now take (OEM) motion server state into account when accepting or executing goals.

If the motion server reports (via the state server) it cannot currently execute motion, either because of a controller error or e-stop, or because a required program is not running on the (OEM) controller, new goals are rejected and currently executing goals are cancelled.

Cancelled trajectories should not resume executing once the (OEM) controller error state is cleared. This will however depend on a cooperating motion server (ie: it must manage its queues properly and watch for and act on errors on the OEM controller).

Cancellations and rejections are accompanied by appropriate human error messages, giving the reason for why a goal was cancelled or rejected. Unfortunately this is mostly only human readable as the FollowJointTrajectory action doesn't currently define appropriately expressive result codes.